+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Improve ZIP reader performance and precision when using `reader-zip` feature
 - Reduce PDF reader memory usage and increase accuracy when using `reader-pdf` feature
 
+## New formats support
+
+- Picture Exchange (PCX)
+
 # Version 0.22.0 (2023-11-04)
 
 ## Internal changes